Separate TestUIReportTests into its own file; fix warning as error in travis
