+++ src/NUnitFramework/nunitlite.tests/nunitlite.tests-4.0.csproj
@@ -77,6 +77,7 @@
     <Compile Include="TeamCityEventListenerTests.cs" />
     <Compile Include="TestNameParserTests.cs" />
     <Compile Include="TestSelectionParserTests.cs" />
+    <Compile Include="TextUIReportTests.cs" />
     <Compile Include="TextUITests.cs" />
     <Compile Include="TokenizerTests.cs" />
   </ItemGroup>
