+++ src/NUnitFramework/nunitlite.tests/nunitlite.tests-4.5.csproj
@@ -80,6 +80,7 @@
     <Compile Include="TeamCityEventListenerTests.cs" />
     <Compile Include="TestNameParserTests.cs" />
     <Compile Include="TestSelectionParserTests.cs" />
+    <Compile Include="TextUIReportTests.cs" />
     <Compile Include="TextUITests.cs" />
     <Compile Include="TokenizerTests.cs" />
   </ItemGroup>
