+++ src/NUnitFramework/nunitlite.tests/nunitlite.tests-portable.csproj
@@ -86,6 +86,7 @@
     <Compile Include="TeamCityEventListenerTests.cs" />
     <Compile Include="TestNameParserTests.cs" />
     <Compile Include="TestSelectionParserTests.cs" />
+    <Compile Include="TextUIReportTests.cs" />
     <Compile Include="TextUITests.cs" />
     <Compile Include="TokenizerTests.cs" />
   </ItemGroup>
